fix(suite): remove circular dependency in TransactionsGraph
- Extracted CryptoGraphProps and related types to separate types.ts file
- GraphTooltipAccount now imports types directly instead of from TransactionsGraph
- This breaks the circular import between TransactionsGraph and GraphTooltipAccount
- Resolves: packages/suite/src/components/suite/graph/TransactionsGraph/TransactionsGraph.tsx > packages/suite/src/components/suite/graph/TransactionsGraph/GraphTooltipAccount.tsx
