+++ packages/suite/src/components/suite/graph/TransactionsGraph/GraphTooltipAccount.tsx
@@ -11,7 +11,7 @@ import { FormattedCryptoAmount } from 'src/components/suite/FormattedCryptoAmoun
 import { CommonAggregatedHistory, GraphRange } from 'src/types/wallet/graph';
 
 import { GraphTooltipBase } from './GraphTooltipBase';
-import type { CryptoGraphProps } from './TransactionsGraph';
+import type { CryptoGraphProps } from './types';
 
 const formatAmount = (
     amount: string | undefined,
